Move test code to a module path that matches src
The DagBag class was moved to a module under dag_processing/dagbag.py
but the test code still remained under a path matching the old module
path (which still exists, with some other classes in it).
